[API-30] Picture added to Team attributes
